changed thoughts, added bookshelf.
i made thoughts 'dynamic', so it doesnt open to a whole new page for lain. but i made a orginazational mess of things bc i copied all the photos back into that file. next to do is to change the paths in the code so they go back a folder and reach into assets. also made bookshelf work, just need to figure out what to do with it, where to put it. also still need to add a guestbook, and still need to add a go back button for all pages. thanks